// enums/classnames.enum.ts
enum ClassName {
  INPUT = 'mt-1 w-full rounded-lg border px-4 py-2 text-sm border-gray-300 focus:outline-none focus:ring-1 focus:ring-accent-500',
  LABEL = 'block text-sm font-medium text-gray-700',
  BUTTON = 'flex items-center gap-1 text-white px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_OUTLINE = 'flex items-center gap-1 text-gray-700 border border-gray-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_DANGER = 'flex items-center gap-1 text-white bg-red-500/90 hover:bg-red-500 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_SUCCESS = 'flex items-center gap-1 text-white bg-green-500/90 hover:bg-green-500 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_WARNING = 'flex items-center gap-1 text-white bg-yellow-500/90 hover:bg-yellow-500 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_INFO = 'flex items-center gap-1 text-white bg-blue-500/90 hover:bg-blue-500 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_DISABLED = 'flex items-center gap-1 text-gray-400 bg-gray-200 cursor-not-allowed px-3 py-2 rounded-md transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK = 'flex items-center gap-1 text-blue-500 hover:text-blue-700 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_OUTLINE = 'flex items-center gap-1 text-gray-700 hover:text-gray-900 border border-gray-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_DANGER = 'flex items-center gap-1 text-red-500 hover:text-red-700 border border-red-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_SUCCESS = 'flex items-center gap-1 text-green-500 hover:text-green-700 border border-green-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_WARNING = 'flex items-center gap-1 text-yellow-500 hover:text-yellow-700 border border-yellow-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_INFO = 'flex items-center gap-1 text-blue-500 hover:text-blue-700 border border-blue-300 px-3 py-2 rounded-md cursor-pointer transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
  BUTTON_LINK_DISABLED = 'flex items-center gap-1 text-gray-400 bg-gray-200 cursor-not-allowed border border-gray-300 px-3 py-2 rounded-md transition duration-200 ease-in-out disabled:opacity-50 disabled:cursor-not-allowed',
}

export { ClassName };